Fixed SRAM issue

+++ v3-my-bedroom/e_global_variables.h
@@ -6,7 +6,7 @@ float temp_vout;
 float aRef_value = 5.0; // Normal: 5.0, Temperature Reading: 1.1, Transition: 0.0
 
 Servo lightAOn_servo, lightBOn_servo, lightAOff_servo, lightBOff_servo;
-long servo_reset_time = 2000L;
-long servo_detach_time = 1000L;
+int servo_reset_time = 2000;
+int servo_detach_time = 1000;
 
-int isLightAOn, isLightBOn;
+byte isLightAOn, isLightBOn;
